requirements.txt
discord.py
meteofrance_api